% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cabone.R
\name{amt_denos}
\alias{amt_denos}
\title{Convert denosumab doses}
\usage{
amt_denos(x)
}
\arguments{
\item{x}{denosumab dose in milligrams}
}
\value{
denosumab dose in mmol
}
\description{
Convert denosumab doses
}
