% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cabone.R
\name{amt_scler}
\alias{amt_scler}
\title{Convert sclerostin doses}
\usage{
amt_scler(x)
}
\arguments{
\item{x}{sclerostin dose in milligrams}
}
\value{
sclerostin dose in nmol
}
\description{
Convert sclerostin doses
}
