% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cabone.R
\name{amt_teri}
\alias{amt_teri}
\title{Convert teriparatide doses}
\usage{
amt_teri(x)
}
\arguments{
\item{x}{teriparatide dose in micrograms}
}
\value{
teriparatide dose in
}
\description{
Convert teriparatide doses
}
\examples{

amt_teri(20)

}
