% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cabone.R
\name{.colSet1}
\alias{.colSet1}
\alias{.colSet2}
\title{Plotting helper functions.}
\usage{
.colSet1(...)

.colSet2(...)
}
\arguments{
\item{...}{passed to \code{ggplot::scale_color_brewer}}
}
\description{
Plotting helper functions.
}
