% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cabone.R
\name{create_scler_plots}
\alias{create_scler_plots}
\title{This is a function to re-create plots from Eudy, et al.(2015), CPT:PSP; Figure 3 ##}
\usage{
create_scler_plots(endpoint)
}
\arguments{
\item{endpoint}{## can be P1NP, CTX, lsBMD, or thBMD}
}
\description{
This is a function to re-create plots from Eudy, et al.(2015), CPT:PSP; Figure 3 ##
}
