% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cabone.R
\name{sim_2h}
\alias{sim_2h}
\title{Simulate secondary hyperparathyroidism}
\usage{
sim_2h(GFRdelta = 84, GFRtau = 10, delta = 24,
  request = "GFR,CaC,PTHpm,OC,ECCPhos", cfb = TRUE)
}
\arguments{
\item{GFRdelta}{change in GFR from baseline value}

\item{GFRtau}{time interval in years over which GFR changes}

\item{delta}{simulation time grid in hours}

\item{request}{outputs to request}

\item{cfb}{if \code{TRUE} results are returned as fractional change from basline}
}
\description{
Simulate secondary hyperparathyroidism
}
