% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cabone.R
\name{sim_combo_arms}
\alias{sim_combo_arms}
\title{This is a function to simulate combination arms (ACOP 2015 poster) ##}
\usage{
sim_combo_arms()
}
\description{
This is a function to simulate combination arms (ACOP 2015 poster) ##
}
\examples{

\dontrun{
  require(ggplot2)
  
  sims <- sim_combo_arms()
  
  ggplot(sims, aes(time,lsBMD,col=regimen)) + 
    geom_line(lwd=1) + theme(legend.position="top")
}

}
