% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cabone.R
\name{sim_scler}
\alias{sim_scler}
\title{Simulate with sclerostin dosing}
\usage{
sim_scler(dose = 210, ii = 1 * 24 * 28, dur = 12, delta = 24 * 28,
  request = "SOSTCP, lsBMDsimSCLER, OBchange")
}
\arguments{
\item{dose}{sclerostin dose in milligrams}

\item{ii}{dosing interval in months}

\item{dur}{number of doses to simulate}

\item{delta}{simulation time grid}

\item{request}{outputs to request}
}
\description{
Simulate with sclerostin dosing
}
