% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cabone.R
\name{sim_scler_data}
\alias{sim_scler_data}
\title{This is a function to re-create plots from Eudy, et al.(2015), CPT:PSP; Figure 3 ##}
\usage{
sim_scler_data()
}
\description{
This is a function to re-create plots from Eudy, et al.(2015), CPT:PSP; Figure 3 ##
}
\examples{

sims <- sim_scler_data()

\dontrun{
  require(ggplot2)

  ggplot(data=sims, aes(time,P1NPsim)) + 
    geom_line() + facet_wrap(~label)

  ggplot(data=sims, aes(time,CTXsim)) + 
    geom_line() + facet_wrap(~label)
  
  ggplot(data=sims, aes(time,lsBMDsimSCLER)) + 
    geom_line() + facet_wrap(~label)
  
  ggplot(data=sims, aes(time,thBMDsimSCLER)) + 
    geom_line() + facet_wrap(~label)
   
}

}
