% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cabone.R
\name{sim_teri}
\alias{sim_teri}
\title{Simulate with teriparatide dosing}
\usage{
sim_teri(dose = 20, ii = 24, dur = 27, delta = 0.1,
  request = "PTHpm,CaC")
}
\arguments{
\item{dose}{teriparatide dose in micrograms}

\item{ii}{dosing interval in hours}

\item{dur}{number of doses to simulate}

\item{delta}{simulation time grid}

\item{request}{outputs to request}
}
\description{
Simulate with teriparatide dosing
}
