% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cabone.R
\name{cabone}
\alias{cabone}
\alias{cabone_export}
\title{Calcium / bone homeostatis model.}
\usage{
cabone(...)

cabone_export(file = tempfile(fileext = ".cpp"), overwrite = FALSE)
}
\arguments{
\item{...}{passed to update}

\item{file}{export file name}

\item{overwrite}{passed to \code{\link{writeLines}}}
}
\description{
Calcium / bone homeostatis model.
}
