% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cabone.R
\name{sim_combo_arms}
\alias{sim_combo_arms}
\title{This is a function to simulate combination arms (ACOP 2015 poster) ##}
\usage{
sim_combo_arms()
}
\value{
plot of combination arms
}
\description{
This is a function to simulate combination arms (ACOP 2015 poster) ##
}
