% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cabone.R
\name{sim_denos}
\alias{sim_denos}
\title{Simulate with denosumab dosing}
\usage{
sim_denos(dose = 60, ii = 6, dur = 3, delta = 4,
  request = "DENCP,DENMOL,BMDlsDENchange", tscale = 1/(24 * 28))
}
\arguments{
\item{dose}{denosumab dose in milligrams}

\item{ii}{dosing interval in months}

\item{dur}{number of doses to simulate}

\item{delta}{simulation time grid in hours}

\item{request}{outputs to request}

\item{tscale}{factor for rescaling time in simulated output}
}
\description{
Simulate with denosumab dosing
}
\examples{

out <- sim_denos(dose=c(10,60,210), dur=6)

plot(out, log(DENCP) + BMDlsDENchange ~ time, xlab="Time (months)")

}
